refactor: adjust build layout to avoid symlink
