+++ bin/vite.js
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../dist/cli')
+require('../dist/node/cli')
