--- src/client/hmrPayload.ts
@@ -1,1 +1,0 @@
-../node/hmrPayload.ts
